mito2
Module access_layer

Structs

AccessLayer

A layer to access SST files under the same directory.

RegionFilePathFactory 🔒

Path provider that builds paths in region storage path.

SstWriteRequest 🔒

Contents to build a SST.

WriteCachePathProvider 🔒

Path provider that builds paths in local write cache.

Enums

OperationType 🔒

OperationType represents the origin of the SstWriteRequest.

Traits

FilePathProvider

Path provider for SST file and index file.

Functions

clean_dir 🔒

Clean the directory.

new_fs_cache_store 🔒

Type Aliases

AccessLayerRef

SstInfoArray

SST write results.