mito2

Module compaction

Modules

buckets 🔒

compactor

memory_manager

picker

reader 🔒

run

This file contains code to find sorted runs in a set if ranged items and along with the best way to merge these items to satisfy the desired run count.

scheduler 🔒

task 🔒

twcs 🔒

window 🔒

Structs

CompactionOutput

CompactionRequest

Region compaction request.

SerializedCompactionOutput

SerializedCompactionOutput is a serialized version of CompactionOutput by replacing FileHandle with FileMeta.

Functions

find_dynamic_options 🔒

Finds compaction options and TTL together with a single metadata fetch to reduce RTT.