mito2::compaction
Module run

This file contains code to find sorted runs in a set if ranged items and along with the best way to merge these items to satisfy the desired run count.

Structs

MergeItems ğŸ”’

SortedRun ğŸ”’

A set of files with non-overlapping time ranges.

Traits

Item ğŸ”’

Trait for items to merge.

Ranged ğŸ”’

Trait for any items with specific range.

Functions

find_sorted_runs ğŸ”’

Finds sorted runs in given items.

merge_all_runs ğŸ”’

reduce_runs ğŸ”’

Reduces the num of runs to given target and returns items to merge. The time complexity of this function is C_{k}_{runs.len()} where k=runs.len()-target+1.

sort_ranged_items ğŸ”’