mito2
Module config

Configurations.

Structs

BloomFilterConfig

Configuration options for the bloom filter.

FulltextIndexConfig

Configuration options for the full-text index.

IndexConfig

InvertedIndexConfig

Configuration options for the inverted index.

MitoConfig

Configuration for MitoEngine.

Enums

MemoryThreshold

Memory threshold for performing certain actions.

Mode

Operational mode for certain actions.

Constants

DEFAULT_SCAN_CHANNEL_SIZE 🔒

Default channel size for parallel scan task.

FETCH_OPTION_TIMEOUT 🔒

Fetch option timeout

GLOBAL_WRITE_BUFFER_SIZE_FACTOR 🔒

INDEX_CONTENT_CACHE_SIZE_FACTOR 🔒

Use 1/INDEX_CONTENT_CACHE_SIZE_FACTOR of OS memory size for inverted index file content cache by default.

INDEX_CREATE_MEM_THRESHOLD_FACTOR 🔒

Use 1/INDEX_CREATE_MEM_THRESHOLD_FACTOR of OS memory size as mem threshold for creating index

MEM_CACHE_SIZE_FACTOR 🔒

Use 1/MEM_CACHE_SIZE_FACTOR of OS memory size as mem cache size in default mode

MULTIPART_UPLOAD_MINIMUM_SIZE 🔒

PAGE_CACHE_SIZE_FACTOR 🔒

Use 1/PAGE_CACHE_SIZE_FACTOR of OS memory size as page cache size in default mode

SST_META_CACHE_SIZE_FACTOR 🔒

Use 1/SST_META_CACHE_SIZE_FACTOR of OS memory size as SST meta cache size in default mode

Functions

divide_num_cpus 🔒

Divide cpu num by a non-zero divisor and returns at least 1.