mito2::engine

Struct MitoEngine

pub struct MitoEngine {
    inner: Arc<EngineInner>,
}

Region engine implementation for timeseries data.

Fields
inner: Arc<EngineInner>
Implementations

impl MitoEngine

pub async fn new<S: LogStore>( data_home: &str, config: MitoConfig, log_store: Arc<S>, object_store_manager: ObjectStoreManagerRef, schema_metadata_manager: SchemaMetadataManagerRef, plugins: Plugins, ) -> Result<MitoEngine>

Returns a new MitoEngine with specific config, log_store and object_store.

pub fn is_region_exists(&self, region_id: RegionId) -> bool

Returns true if the specific region exists.

pub fn is_region_opening(&self, region_id: RegionId) -> bool

Returns true if the specific region exists.

pub fn get_region_statistic( &self, region_id: RegionId, ) -> Option<RegionStatistic>

Returns the region disk/memory statistic.

pub fn get_primary_key_encoding( &self, region_id: RegionId, ) -> Option<PrimaryKeyEncoding>

Returns primary key encoding of the region.

pub async fn scan_to_stream( &self, region_id: RegionId, request: ScanRequest, ) -> Result<SendableRecordBatchStream, BoxedError>

Handle substrait query and return a stream of record batches

Notice that the output stream’s ordering is not guranateed. If order matter, please use [scanner] to build a Scanner to consume.

fn scanner(&self, region_id: RegionId, request: ScanRequest) -> Result<Scanner>

Returns a scanner to scan for request.

fn scan_region( &self, region_id: RegionId, request: ScanRequest, ) -> Result<ScanRegion>

Scans a region.

pub async fn edit_region( &self, region_id: RegionId, edit: RegionEdit, ) -> Result<()>

Edit region’s metadata by RegionEdit directly. Use with care. Now we only allow adding files to region (the RegionEdit struct can only contain a non-empty “files_to_add” field). Other region editing intention will result in an “invalid request” error. Also note that if a region is to be edited directly, we MUST not write data to it thereafter.

fn encode_manifest_info_to_extensions( region_id: &RegionId, manifest_info: RegionManifestInfo, extensions: &mut HashMap<String, Vec<u8>>, ) -> Result<()>