mito2::error

Enum Error

pub enum Error {

    DataTypeMismatch {
        source: Error,
        location: Location,
    },
    External {
        source: BoxedError,
        context: String,
        location: Location,
    },
    EncodeSparsePrimaryKey {
        reason: String,
        location: Location,
    },
    OpenDal {
        location: Location,
        error: Error,
    },
    CompressObject {
        compress_type: CompressionType,
        path: String,
        error: Error,
    },
    DecompressObject {
        compress_type: CompressionType,
        path: String,
        error: Error,
    },
    SerdeJson {
        location: Location,
        error: Error,
    },
    SerializeColumnMetadata {
        error: Error,
        location: Location,
    },
    SerializeManifest {
        region_id: RegionId,
        error: Error,
        location: Location,
    },
    InvalidScanIndex {
        start: ManifestVersion,
        end: ManifestVersion,
        location: Location,
    },
    Utf8 {
        location: Location,
        error: Utf8Error,
    },
    RegionMetadataNotFound {
        location: Location,
    },
    Join {
        error: JoinError,
        location: Location,
    },
    WorkerStopped {
        id: u32,
        location: Location,
    },
    Recv {
        error: RecvError,
        location: Location,
    },
    InvalidMeta {
        reason: String,
        location: Location,
    },
    InvalidMetadata {
        source: MetadataError,
        location: Location,
    },
    NewRecordBatch {
        location: Location,
        error: ArrowError,
    },
    ReadParquet {
        path: String,
        error: ParquetError,
        location: Location,
    },
    WriteParquet {
        error: ParquetError,
        location: Location,
    },
    RegionNotFound {
        region_id: RegionId,
        location: Location,
    },
    ObjectStoreNotFound {
        object_store: String,
        location: Location,
    },
    RegionCorrupted {
        region_id: RegionId,
        reason: String,
        location: Location,
    },
    InvalidRequest {
        region_id: RegionId,
        reason: String,
        location: Location,
    },
    MissingOldManifest {
        region_id: RegionId,
        location: Location,
    },
    MissingNewManifest {
        region_id: RegionId,
        location: Location,
    },
    MissingManifest {
        region_id: RegionId,
        location: Location,
    },
    InconsistentFile {
        file_id: FileId,
        reason: String,
        location: Location,
    },
    FilesLost {
        old_count: usize,
        new_count: usize,
        location: Location,
    },
    NoOldManifests {
        location: Location,
    },
    FetchManifests {
        location: Location,
        source: BoxedError,
    },
    MissingPartitionExpr {
        region_id: RegionId,
        location: Location,
    },
    SerializePartitionExpr {
        source: Error,
        location: Location,
    },
    ConvertColumnDataType {
        reason: String,
        source: Error,
        location: Location,
    },
    FillDefault {
        region_id: RegionId,
    },
    CreateDefault {
        region_id: RegionId,
        column: String,
        source: Error,
        location: Location,
    },
    BuildEntry {
        region_id: RegionId,
        location: Location,
        source: BoxedError,
    },
    WriteWal {
        location: Location,
        source: BoxedError,
    },
    ReadWal {
        provider: Provider,
        location: Location,
        source: BoxedError,
    },
    DecodeWal {
        region_id: RegionId,
        location: Location,
        error: DecodeError,
    },
    DeleteWal {
        region_id: RegionId,
        location: Location,
        source: BoxedError,
    },
    WriteGroup {
        source: Arc<Error>,
    },
    InvalidParquet {
        file: String,
        reason: String,
        location: Location,
    },
    InvalidBatch {
        reason: String,
        location: Location,
    },
    InvalidRecordBatch {
        reason: String,
        location: Location,
    },
    InvalidWalReadRequest {
        reason: String,
        location: Location,
    },
    ConvertVector {
        location: Location,
        source: Error,
    },
    ComputeArrow {
        location: Location,
        error: ArrowError,
    },
    EvalPartitionFilter {
        location: Location,
        error: DataFusionError,
    },
    ComputeVector {
        location: Location,
        source: Error,
    },
    PrimaryKeyLengthMismatch {
        expect: usize,
        actual: usize,
        location: Location,
    },
    InvalidSender {
        location: Location,
    },
    InvalidSchedulerState {
        location: Location,
    },
    StopScheduler {
        error: JoinError,
        location: Location,
    },
    DeleteSst {
        file_id: FileId,
        error: Error,
        location: Location,
    },
    DeleteIndex {
        file_id: FileId,
        error: Error,
        location: Location,
    },
    FlushRegion {
        region_id: RegionId,
        source: Arc<Error>,
        location: Location,
    },
    RegionDropped {
        region_id: RegionId,
        location: Location,
    },
    RegionClosed {
        region_id: RegionId,
        location: Location,
    },
    RegionTruncated {
        region_id: RegionId,
        location: Location,
    },
    RejectWrite {
        region_id: RegionId,
        location: Location,
    },
    CompactRegion {
        region_id: RegionId,
        source: Arc<Error>,
        location: Location,
    },
    CompatReader {
        region_id: RegionId,
        reason: String,
        location: Location,
    },
    InvalidRegionRequest {
        source: MetadataError,
        location: Location,
    },
    UpdateManifest {
        region_id: RegionId,
        state: RegionRoleState,
        location: Location,
    },
    RegionState {
        region_id: RegionId,
        state: RegionRoleState,
        expect: RegionRoleState,
        location: Location,
    },
    PartitionExprVersionMismatch {
        region_id: RegionId,
        request_version: u64,
        expected_version: u64,
        location: Location,
    },
    JsonOptions {
        error: Error,
        location: Location,
    },
    EmptyRegionDir {
        region_id: RegionId,
        region_dir: String,
        location: Location,
    },
    EmptyManifestDir {
        manifest_dir: String,
        location: Location,
    },
    ArrowReader {
        path: String,
        error: ArrowError,
        location: Location,
    },
    ColumnNotFound {
        column: String,
        location: Location,
    },
    BuildIndexApplier {
        source: Error,
        location: Location,
    },
    BuildIndexAsync {
        region_id: RegionId,
        source: Arc<Error>,
        location: Location,
    },
    ConvertValue {
        source: Error,
        location: Location,
    },
    ApplyInvertedIndex {
        source: Error,
        location: Location,
    },
    ApplyBloomFilterIndex {
        source: Error,
        location: Location,
    },
    PushIndexValue {
        source: Error,
        location: Location,
    },
    IndexFinish {
        source: Error,
        location: Location,
    },
    OperateAbortedIndex {
        location: Location,
    },
    PuffinReadBlob {
        source: Error,
        location: Location,
    },
    PuffinAddBlob {
        source: Error,
        location: Location,
    },
    CleanDir {
        dir: String,
        error: Error,
        location: Location,
    },
    InvalidConfig {
        reason: String,
        location: Location,
    },
    StaleLogEntry {
        region_id: RegionId,
        flushed_entry_id: u64,
        unexpected_entry_id: u64,
    },
    Download {
        region_id: RegionId,
        file_id: FileId,
        file_type: FileType,
        error: Error,
        location: Location,
    },
    Upload {
        region_id: RegionId,
        file_id: FileId,
        file_type: FileType,
        error: Error,
        location: Location,
    },
    CreateDir {
        dir: String,
        error: Error,
    },
    RecordBatch {
        source: Error,
        location: Location,
    },
    BiErrors {
        first: Box<Error>,
        second: Box<Error>,
        location: Location,
    },
    IndexEncodeNull {
        location: Location,
    },
    EncodeMemtable {
        error: ParquetError,
        location: Location,
    },
    PartitionOutOfRange {
        given: usize,
        all: usize,
        location: Location,
    },
    ReadDataPart {
        location: Location,
        error: ParquetError,
    },
    DecodeArrowRowGroup {
        error: ArrowError,
        location: Location,
    },
    InvalidRegionOptions {
        reason: String,
        location: Location,
    },
    ChecksumMismatch {
        actual: u32,
        expected: u32,
    },
    NoCheckpoint {
        region_id: RegionId,
        last_version: ManifestVersion,
        location: Location,
    },
    NoManifests {
        region_id: RegionId,
        start_version: ManifestVersion,
        end_version: ManifestVersion,
        last_version: ManifestVersion,
        location: Location,
    },
    InstallManifestTo {
        region_id: RegionId,
        target_version: ManifestVersion,
        available_version: ManifestVersion,
        location: Location,
        last_version: ManifestVersion,
    },
    RegionStopped {
        region_id: RegionId,
        location: Location,
    },
    TimeRangePredicateOverflow {
        timestamp: Timestamp,
        unit: TimeUnit,
        location: Location,
    },
    OpenRegion {
        location: Location,
        source: Arc<Error>,
    },
    ParseJobId {
        location: Location,
        error: Error,
    },
    UnsupportedOperation {
        err_msg: String,
        location: Location,
    },
    RemoteCompaction {
        region_id: RegionId,
        job_id: Option<JobId>,
        reason: String,
        location: Location,
    },
    PuffinInitStager {
        source: Error,
        location: Location,
    },
    PuffinPurgeStager {
        source: Error,
        location: Location,
    },
    PuffinBuildReader {
        source: Error,
        location: Location,
    },
    IndexOptions {
        location: Location,
        source: Error,
        column_name: String,
    },
    CreateFulltextCreator {
        source: Error,
        location: Location,
    },
    CastVector {
        location: Location,
        from: ConcreteDataType,
        to: ConcreteDataType,
        source: Error,
    },
    FulltextPushText {
        source: Error,
        location: Location,
    },
    FulltextFinish {
        source: Error,
        location: Location,
    },
    ApplyFulltextIndex {
        source: Error,
        location: Location,
    },
    StatsNotPresent {
        file_path: String,
        location: Location,
    },
    DecodeStats {
        file_path: String,
        location: Location,
    },
    RegionBusy {
        region_id: RegionId,
        location: Location,
    },
    GetSchemaMetadata {
        source: Error,
        location: Location,
    },
    Timeout {
        error: Elapsed,
        location: Location,
    },
    Metadata {
        error: Error,
        location: Location,
    },
    PushBloomFilterValue {
        source: Error,
        location: Location,
    },
    BloomFilterFinish {
        source: Error,
        location: Location,
    },
    ManualCompactionOverride {},
    CompactionMemoryExhausted {
        region_id: RegionId,
        policy: String,
        source: Error,
        location: Location,
    },
    IncompatibleWalProviderChange {
        global: String,
        region: String,
    },
    MitoManifestInfo {
        location: Location,
    },
    ScanSeries {
        location: Location,
        source: Arc<Error>,
    },
    ScanMultiTimes {
        partition: usize,
        location: Location,
    },
    InvalidPartitionExpr {
        expr: String,
        location: Location,
        source: Error,
    },
    ConvertBulkWalEntry {
        location: Location,
        source: Error,
    },
    Encode {
        location: Location,
        source: Error,
    },
    Decode {
        location: Location,
        source: Error,
    },
    Unexpected {
        reason: String,
        location: Location,
    },
    InconsistentTimestampLength {
        expected: usize,
        actual: usize,
        location: Location,
    },
    TooManyFilesToRead {
        actual: usize,
        max: usize,
        location: Location,
    },
    DurationOutOfRange {
        input: Duration,
        error: OutOfRangeError,
        location: Location,
    },
    TooManyGcJobs {
        location: Location,
    },
    StagingPartitionExprMismatch {
        manifest_expr: Option<String>,
        request_expr: String,
        location: Location,
    },
    InvalidSourceAndTargetRegion {
        source_region_id: RegionId,
        target_region_id: RegionId,
        location: Location,
    },
    PruneFile {
        source: Arc<Error>,
        location: Location,
    },

}

Variants

DataTypeMismatch

Fields

source: Error

location: Location

External

Fields

source: BoxedError

context: String

location: Location

EncodeSparsePrimaryKey

Fields

reason: String

location: Location

OpenDal

Fields

location: Location

error: Error

CompressObject

Fields

compress_type: CompressionType

path: String

error: Error

DecompressObject

Fields

compress_type: CompressionType

path: String

error: Error

SerdeJson

Fields

location: Location

error: Error

SerializeColumnMetadata

Fields

error: Error

location: Location

SerializeManifest

Fields

region_id: RegionId

error: Error

location: Location

InvalidScanIndex

Fields

start: ManifestVersion

end: ManifestVersion

location: Location

Utf8

Fields

location: Location

error: Utf8Error

RegionMetadataNotFound

Fields

location: Location

Join

Fields

error: JoinError

location: Location

WorkerStopped

Fields

id: u32

location: Location

Recv

Fields

error: RecvError

location: Location

InvalidMeta

Fields

reason: String

location: Location

InvalidMetadata

Fields

source: MetadataError

location: Location

NewRecordBatch

Fields

location: Location

error: ArrowError

ReadParquet

Fields

path: String

error: ParquetError

location: Location

WriteParquet

Fields

error: ParquetError

location: Location

RegionNotFound

Fields

region_id: RegionId

location: Location

ObjectStoreNotFound

Fields

object_store: String

location: Location

RegionCorrupted

Fields

region_id: RegionId

reason: String

location: Location

InvalidRequest

Fields

region_id: RegionId

reason: String

location: Location

MissingOldManifest

Fields

region_id: RegionId

location: Location

MissingNewManifest

Fields

region_id: RegionId

location: Location

MissingManifest

Fields

region_id: RegionId

location: Location

InconsistentFile

Fields

file_id: FileId

reason: String

location: Location

FilesLost

Fields

old_count: usize

new_count: usize

location: Location

NoOldManifests

Fields

location: Location

FetchManifests

Fields

location: Location

source: BoxedError

MissingPartitionExpr

Fields

region_id: RegionId

location: Location

SerializePartitionExpr

Fields

source: Error

location: Location

ConvertColumnDataType

Fields

reason: String

source: Error

location: Location

FillDefault

An error type to indicate that schema is changed and we need to fill default values again.

Fields

region_id: RegionId

CreateDefault

Fields

region_id: RegionId

column: String

source: Error

location: Location

BuildEntry

Fields

region_id: RegionId

location: Location

source: BoxedError

WriteWal

Fields

location: Location

source: BoxedError

ReadWal

Fields

provider: Provider

location: Location

source: BoxedError

DecodeWal

Fields

region_id: RegionId

location: Location

error: DecodeError

DeleteWal

Fields

region_id: RegionId

location: Location

source: BoxedError

WriteGroup

Fields

source: Arc<Error>

InvalidParquet

Fields

file: String

reason: String

location: Location

InvalidBatch

Fields

reason: String

location: Location

InvalidRecordBatch

Fields

reason: String

location: Location

InvalidWalReadRequest

Fields

reason: String

location: Location

ConvertVector

Fields

location: Location

source: Error

ComputeArrow

Fields

location: Location

error: ArrowError

EvalPartitionFilter

Fields

location: Location

error: DataFusionError

ComputeVector

Fields

location: Location

source: Error

PrimaryKeyLengthMismatch

Fields

expect: usize

actual: usize

location: Location

InvalidSender

Fields

location: Location

InvalidSchedulerState

Fields

location: Location

StopScheduler

Fields

error: JoinError

location: Location

DeleteSst

Fields

file_id: FileId

error: Error

location: Location

DeleteIndex

Fields

file_id: FileId

error: Error

location: Location

FlushRegion

Fields

region_id: RegionId

source: Arc<Error>

location: Location

RegionDropped

Fields

region_id: RegionId

location: Location

RegionClosed

Fields

region_id: RegionId

location: Location

RegionTruncated

Fields

region_id: RegionId

location: Location

RejectWrite

Fields

region_id: RegionId

location: Location

CompactRegion

Fields

region_id: RegionId

source: Arc<Error>

location: Location

CompatReader

Fields

region_id: RegionId

reason: String

location: Location

InvalidRegionRequest

Fields

source: MetadataError

location: Location

UpdateManifest

Fields

region_id: RegionId

state: RegionRoleState

location: Location

RegionState

Fields

region_id: RegionId

state: RegionRoleState

expect: RegionRoleState

location: Location

PartitionExprVersionMismatch

Fields

region_id: RegionId

request_version: u64

expected_version: u64

location: Location

JsonOptions

Fields

error: Error

location: Location

EmptyRegionDir

Fields

region_id: RegionId

region_dir: String

location: Location

EmptyManifestDir

Fields

manifest_dir: String

location: Location

ArrowReader

Fields

path: String

error: ArrowError

location: Location

ColumnNotFound

Fields

column: String

location: Location

BuildIndexApplier

Fields

source: Error

location: Location

BuildIndexAsync

Fields

region_id: RegionId

source: Arc<Error>

location: Location

ConvertValue

Fields

source: Error

location: Location

ApplyInvertedIndex

Fields

source: Error

location: Location

ApplyBloomFilterIndex

Fields

source: Error

location: Location

PushIndexValue

Fields

source: Error

location: Location

IndexFinish

Fields

source: Error

location: Location

OperateAbortedIndex

Fields

location: Location

PuffinReadBlob

Fields

source: Error

location: Location

PuffinAddBlob

Fields

source: Error

location: Location

CleanDir

Fields

dir: String

error: Error

location: Location

InvalidConfig

Fields

reason: String

location: Location

StaleLogEntry

Fields

region_id: RegionId

flushed_entry_id: u64

unexpected_entry_id: u64

Download

Fields

region_id: RegionId

file_id: FileId

file_type: FileType

error: Error

location: Location

Upload

Fields

region_id: RegionId

file_id: FileId

file_type: FileType

error: Error

location: Location

CreateDir

Fields

dir: String

error: Error

RecordBatch

Fields

source: Error

location: Location

BiErrors

Fields

first: Box<Error>

second: Box<Error>

location: Location

IndexEncodeNull

Fields

location: Location

EncodeMemtable

Fields

error: ParquetError

location: Location

PartitionOutOfRange

Fields

given: usize

all: usize

location: Location

ReadDataPart

Fields

location: Location

error: ParquetError

DecodeArrowRowGroup

Fields

error: ArrowError

location: Location

InvalidRegionOptions

Fields

reason: String

location: Location

ChecksumMismatch

Fields

actual: u32

expected: u32

NoCheckpoint

Fields

region_id: RegionId

last_version: ManifestVersion

location: Location

NoManifests

Fields

region_id: RegionId

start_version: ManifestVersion

end_version: ManifestVersion

last_version: ManifestVersion

location: Location

InstallManifestTo

Fields

region_id: RegionId

target_version: ManifestVersion

available_version: ManifestVersion

location: Location

last_version: ManifestVersion

RegionStopped

Fields

region_id: RegionId

location: Location

TimeRangePredicateOverflow

Fields

timestamp: Timestamp

unit: TimeUnit

location: Location

OpenRegion

Fields

location: Location

source: Arc<Error>

ParseJobId

Fields

location: Location

error: Error

UnsupportedOperation

Fields

err_msg: String

location: Location

RemoteCompaction

Fields

region_id: RegionId

job_id: Option<JobId>

reason: String

location: Location

PuffinInitStager

Fields

source: Error

location: Location

PuffinPurgeStager

Fields

source: Error

location: Location

PuffinBuildReader

Fields

source: Error

location: Location

IndexOptions

Fields

location: Location

source: Error

column_name: String

CreateFulltextCreator

Fields

source: Error

location: Location

CastVector

Fields

location: Location

from: ConcreteDataType

to: ConcreteDataType

source: Error

FulltextPushText

Fields

source: Error

location: Location

FulltextFinish

Fields

source: Error

location: Location

ApplyFulltextIndex

Fields

source: Error

location: Location

StatsNotPresent

Fields

file_path: String

location: Location

DecodeStats

Fields

file_path: String

location: Location

RegionBusy

Fields

region_id: RegionId

location: Location

GetSchemaMetadata

Fields

source: Error

location: Location

Timeout

Fields

error: Elapsed

location: Location

Metadata

Fields

error: Error

location: Location

PushBloomFilterValue

Fields

source: Error

location: Location

BloomFilterFinish

Fields

source: Error

location: Location

ManualCompactionOverride

CompactionMemoryExhausted

Fields

region_id: RegionId

policy: String

source: Error

location: Location

IncompatibleWalProviderChange

Fields

global: String

region: String

MitoManifestInfo

Fields

location: Location

ScanSeries

Fields

location: Location

source: Arc<Error>

ScanMultiTimes

Fields

partition: usize

location: Location

InvalidPartitionExpr

Fields

expr: String

location: Location

source: Error

ConvertBulkWalEntry

Fields

location: Location

source: Error

Encode

Fields

location: Location

source: Error

Decode

Fields

location: Location

source: Error

Unexpected

Fields

reason: String

location: Location

InconsistentTimestampLength

Fields

expected: usize

actual: usize

location: Location

TooManyFilesToRead

Fields

actual: usize

max: usize

location: Location

DurationOutOfRange

Fields

input: Duration

error: OutOfRangeError

location: Location

TooManyGcJobs

Fields

location: Location

StagingPartitionExprMismatch

Fields

manifest_expr: Option<String>

request_expr: String

location: Location

InvalidSourceAndTargetRegion

Fields

source_region_id: RegionId

target_region_id: RegionId

location: Location

PruneFile

Fields

source: Arc<Error>

location: Location

Implementations

impl Error

pub(crate) fn is_fill_default(&self) -> bool

Returns true if we need to fill default value for a region.

pub(crate) fn is_object_not_found(&self) -> bool

Returns true if the file is not found on the object store.