mito2::gc

Function gen_partition_from_concurrency

fn gen_partition_from_concurrency(concurrency: usize) -> Vec<String>

Generate partition prefixes based on concurrency and assume file names are evenly-distributed uuid string, to evenly distribute files across partitions. For example, if concurrency is 2, partition prefixes will be: [“8”] so it divide uuids into two partitions based on the first character. If concurrency is 32, partition prefixes will be: [“08”, “10”, “18”, “20”, “28”, “30”, “38” …, “f0”, “f8”] if concurrency is 1, it returns an empty vector.