mito2::manifest

Module action

Defines RegionMetaAction related structs and RegionCheckpoint.

Structs

RegionChange

RegionCheckpoint

RegionEdit

RegionManifest

The region manifest data.

RegionManifestBuilder

RegionMetaActionList

RegionRemove

RegionTruncate

Last data truncated in the region.

RemovedFiles

RemovedFilesRecord

A record of removed files in the region manifest. This is used to keep track of files that have been removed from the manifest but may still be kept for a while

Enums

RegionMetaAction

Actions that can be applied to region manifest.

RemovedFile

A removed file, which can be a data file(optional paired with a index file) or an outdated index file.

TruncateKind

The kind of truncate operation.