mito2::memtable::bulk

Module part

Bulk part encoder/decoder.

Structs

BulkPart

A raw bulk part in the memtable.

BulkPartConverter

Converter that converts structs into BulkPart.

BulkPartEncodeMetrics

Metrics for encoding a part.

BulkPartEncoder

BulkPartMeta

EncodedBulkPart

UnorderedPart

A collection of small unordered bulk parts. Used to batch small parts together before merging them into a sorted part.

Enums

PrimaryKeyColumnBuilder 🔒

Primary key column builder for handling strings specially.

Constants

INIT_DICT_VALUE_CAPACITY 🔒

Functions

convert_bulk_part

Converts a BulkPart that is unordered and without encoded primary keys into a BulkPart with the same format as produced by BulkPartConverter.

new_primary_key_column_builders 🔒

record_batch_estimated_size

More accurate estimation of the size of a record batch.

sort_primary_key_record_batch

Sorts the record batch with primary key format.