Module mito2::memtable::partition_tree::data

Data part of a shard.

Structs

DataBatch

Data part batches returns by DataParts::read.

DataBatchRange 🔒

Range of a data batch.

DataBuffer

Buffer for the value part (pk_index, ts, sequence, op_type, field columns) in a shard.

DataBufferReader 🔒

DataBufferReaderBuilder 🔒

DataPartEncoder 🔒

DataPartReader

DataParts

Data parts under a shard.

DataPartsReader

Reader for all parts inside a DataParts.

DataPartsReaderBuilder

InnerKey 🔒

ParquetPart

Parquet-encoded DataPart.

Enums

DataPart

Format of immutable data part.

LazyFieldVector 🔒

LazyMutableVectorBuilder 🔒

Constants

DATA_INIT_CAP 🔒

Initial capacity for the data buffer.

PK_INDEX_COLUMN_NAME 🔒

Functions

build_row_sort_indices_and_columns 🔒

build_rows_to_sort 🔒

drain_data_buffer_to_record_batches 🔒

Converts DataBuffer to record batches, with rows sorted according to pk_weights. dedup: whether to true to remove the duplicated rows inside DataBuffer. replace_pk_index: whether to replace the pk_index values with corresponding pk weight.

memtable_schema_to_encoded_schema 🔒

pk_index_array 🔒

Gets pk_index array from record batch.

search_next_pk_range 🔒

Searches for next pk index, and it’s offset range in a sorted UInt16Array.

timestamp_array_to_i64_slice 🔒