mito2::memtable::partition_tree

Module dict

Key dictionary of a shard.

Structs

DictBlock 🔒

A block in the key dictionary.

DictBuilderReader

Reader to scan the KeyDictBuilder.

KeyBuffer 🔒

Buffer to store unsorted primary keys.

KeyDict

A key dictionary.

KeyDictBuilder

Builder to build a key dictionary.

Constants

MAX_KEYS_PER_BLOCK 🔒

Maximum keys in a DictBlock.

Functions

compute_pk_weights 🔒

Returns pk weights to sort a data part and replaces pk indices.

Type Aliases

KeyDictRef

PkIndexMap 🔒

The key is mcmp-encoded primary keys, while the values are the pk index and optionally sparsely encoded primary keys.