mito2::memtable
Module time_series

Structs

Iter 🔒

Metrics 🔒

Metrics for reading the memtable.

Series 🔒

A Series holds a list of field values of some given primary key.

SeriesSet 🔒

TimeSeriesIterBuilder 🔒

TimeSeriesMemtable

Memtable implementation that groups rows by their primary key.

TimeSeriesMemtableBuilder

Builder to build TimeSeriesMemtable.

ValueBuilder 🔒

ValueBuilder holds all the vector builders for field columns.

Values 🔒

Values holds an immutable vectors of field columns, including sequence and op_type.

Constants

BUILDER_CAPACITY 🔒

Vector builder capacity.

INITIAL_BUILDER_CAPACITY 🔒

Initial vector builder capacity.

Functions

primary_key_schema 🔒

Creates an arrow SchemaRef that only contains primary keys of given region schema

prune_primary_key 🔒

Type Aliases

SeriesRwLockMap 🔒