mito2::memtable
Trait IterBuilder

pub trait IterBuilder: Send + Sync {
    // Required method
    fn build(&self) -> Result<BoxedBatchIterator>;
}

Builder to build an iterator to read the range. The builder should know the projection and the predicate to build the iterator.

Required Methods

fn build(&self) -> Result<BoxedBatchIterator>

Returns the iterator to read the range.