mito2
Module metrics

Structs

CACHE_BYTES

Cache size in bytes.

CACHE_EVICTION

Cache eviction counter, labeled with cache type and eviction reason.

CACHE_HIT

Cache hit counter.

CACHE_MISS

Cache miss counter.

COMPACTION_ELAPSED_TOTAL

Timer of whole compaction task.

COMPACTION_FAILURE_COUNT

Counter of failed compaction task.

COMPACTION_REQUEST_COUNT

Counter of all requested compaction task.

COMPACTION_STAGE_ELAPSED

Timer of different stages in compaction.

FLUSH_BYTES_TOTAL

Histogram of flushed bytes.

FLUSH_ELAPSED

Elapsed time of a flush job.

FLUSH_ERRORS_TOTAL

Counter of scheduled failed flush jobs.

FLUSH_REQUESTS_TOTAL

Counter of scheduled flush requests. Note that the flush scheduler may merge some flush requests.

HANDLE_REQUEST_ELAPSED

Elapsed time to handle requests.

INDEX_APPLY_ELAPSED

Timer of index application.

INDEX_APPLY_MEMORY_USAGE

Gauge of index apply memory usage.

INDEX_CREATE_BYTES_TOTAL

Counter of created index bytes.

INDEX_CREATE_ELAPSED

Timer of index creation.

INDEX_CREATE_MEMORY_USAGE

Gauge of index create memory usage.

INDEX_CREATE_ROWS_TOTAL

Counter of rows indexed.

INDEX_INTERMEDIATE_FLUSH_OP_TOTAL

Counter of flush operations on intermediate files.

INDEX_INTERMEDIATE_READ_BYTES_TOTAL

Counter of read bytes on intermediate files.

INDEX_INTERMEDIATE_READ_OP_TOTAL

Counter of read operations on intermediate files.

INDEX_INTERMEDIATE_SEEK_OP_TOTAL

Counter of seek operations on intermediate files.

INDEX_INTERMEDIATE_WRITE_BYTES_TOTAL

Counter of write bytes on intermediate files.

INDEX_INTERMEDIATE_WRITE_OP_TOTAL

Counter of write operations on intermediate files.

INDEX_IO_BYTES_TOTAL

Counter of r/w bytes on index related IO operations.

INDEX_IO_OP_TOTAL

Counter of r/w operations on index related IO operations, e.g. read, write, seek and flush.

INDEX_PUFFIN_FLUSH_OP_TOTAL

Counter of flush operations on puffin files.

INDEX_PUFFIN_READ_BYTES_TOTAL

Counter of read bytes on puffin files.

INDEX_PUFFIN_READ_OP_TOTAL

Counter of read operations on puffin files.

INDEX_PUFFIN_SEEK_OP_TOTAL

Counter of seek operations on puffin files.

INDEX_PUFFIN_WRITE_BYTES_TOTAL

Counter of write bytes on puffin files.

INDEX_PUFFIN_WRITE_OP_TOTAL

Counter of write operations on puffin files.

MANIFEST_OP_ELAPSED

Elapsed time of manifest operation. Labeled with “op”.

MEMTABLE_DICT_BYTES

Global memtable dictionary size in bytes.

MERGE_FILTER_ROWS_TOTAL

Counter of filtered rows during merge.

PARTITION_TREE_DATA_BUFFER_FREEZE_STAGE_ELAPSED

Partition tree memtable data buffer freeze metrics

PARTITION_TREE_READ_STAGE_ELAPSED

Partition tree memtable read path metrics

PRECISE_FILTER_ROWS_TOTAL

Counter of filtered rows by precise filter.

READ_BATCHES_RETURN

Histogram for the number of batches returned per query.

READ_ROWS_IN_ROW_GROUP_TOTAL

READ_ROWS_RETURN

Histogram for the number of rows returned per query.

READ_ROWS_TOTAL

Counter of rows read from different source.

READ_ROW_GROUPS_TOTAL

Counter of row groups read.

READ_SST_COUNT

Histogram for the number of SSTs to scan per query.

READ_STAGE_ELAPSED

Timer of different stages in query.

READ_STAGE_FETCH_PAGES

REGION_COUNT

Gauge for open regions in each worker.

UPLOAD_BYTES_TOTAL

Upload bytes counter.

WRITE_BUFFER_BYTES

Global write buffer size in bytes.

WRITE_CACHE_DOWNLOAD_BYTES_TOTAL

Download bytes counter in the write cache.

WRITE_CACHE_DOWNLOAD_ELAPSED

Timer of the downloading task in the write cache.

WRITE_REJECT_TOTAL

Counter of rejected write requests.

WRITE_ROWS_TOTAL

Counter of rows to write.

WRITE_STAGE_ELAPSED

Elapsed time of each write stage.

WRITE_STALL_TOTAL

Number of stalled write requests in each worker.

Constants

CACHE_EVICTION_CAUSE 🔒

FILE_TYPE_LABEL

File type label.

FLUSH_REASON

Reason to flush.

STAGE_LABEL

Stage label.

TYPE_LABEL

Type label.

WORKER_LABEL

Region worker id label.