mito2::read

Module compat

Utilities to adapt readers with different schema.

Structs

CompatFields 🔒

Helper to make fields compatible.

CompatPrimaryKey 🔒

Helper to make primary key compatible.

CompatReader

Reader to adapt schema of underlying reader to expected schema.

FlatCompatBatch 🔒

A helper struct to adapt schema of the batch to an expected schema.

FlatCompatPrimaryKey 🔒

Helper to make primary key compatible for flat format.

FlatRewritePrimaryKey 🔒

Helper to rewrite primary key to another encoding for flat format.

PrimaryKeyCompatBatch 🔒

A helper struct to adapt schema of the batch to an expected schema.

RewritePrimaryKey 🔒

Adapter to rewrite primary key.

Enums

CompatBatch 🔒

Helper to adapt schema of the batch to an expected schema.

IndexOrDefault 🔒

Index in source batch or a default value to fill a column.

Functions

has_same_columns_and_pk_encoding 🔒

Returns true if left and right have same columns and primary key encoding.

is_primary_key_same 🔒

Returns true if the actual primary keys is the same as expected.

may_compat_fields 🔒

Creates a CompatFields if needed.

may_compat_primary_key 🔒

Creates a CompatPrimaryKey if needed.

may_rewrite_primary_key 🔒

repeat_vector 🔒

Repeats the vector value to_len times.