mito2::read

Module flat_merge

Structs

BatchBuilder

Provides an API to incrementally build a [RecordBatch] from partitioned [RecordBatch]

BatchCursor 🔒

Keeps track of the current position in a batch

FlatMergeIterator

Iterator to merge multiple sorted iterators into a single sorted iterator.

FlatMergeReader

Iterator to merge multiple sorted iterators into a single sorted iterator.

GenericNode 🔒

A sync node in the merge iterator.

MergeAlgo 🔒

Common algorithm of merging sorted batches from multiple nodes.

RowCursor 🔒

Cursor to a row in the [RecordBatch].

SortColumns 🔒

Columns to compare for a [RecordBatch].

Traits

NodeCmp 🔒

A comparable node of the heap.

Type Aliases

IterNode 🔒

StreamNode 🔒