mito2::read::flat_merge

Trait NodeCmp

trait NodeCmp: Eq + Ord {
    // Required methods
    fn is_eof(&self) -> bool;
    fn is_behind(&self, other: &Self) -> bool;
}

A comparable node of the heap.

Required Methods

fn is_eof(&self) -> bool

Returns whether the node still has batch to read.

fn is_behind(&self, other: &Self) -> bool

Returns true if the key range of current batch in self is behind (exclusive) current batch in other.

Panics

Panics if either self or other is EOF.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.