mito2::read
Module projection

Utilities for projection.

Structs

ProjectionMapper

Handles projection and converts a projected Batch to a projected [RecordBatch].

Enums

BatchIndex 🔒

Index of a vector in a Batch.

Constants

MAX_VECTOR_LENGTH_TO_CACHE 🔒

Only cache vector when its length <= this value.

Functions

new_repeated_vector 🔒

Returns a vector with repeated values.

repeated_vector_with_cache 🔒

Gets a vector with repeated values from specific cache or creates a new one.