Module mito2::read::range

Structs for partition ranges.

Structs

RangeMeta 🔒

Meta data of a partition range. If the scanner is [UnorderedScan], each meta only has one row group or memtable. If the scanner is [SeqScan], each meta may have multiple row groups and memtables.

RowGroupIndex 🔒

Index to access a row group.

Constants

ALL_ROW_GROUPS 🔒

Functions

group_ranges_for_seq_scan 🔒

Groups ranges by time range. It assumes each input range only contains a file or a memtable.

maybe_split_ranges_for_seq_scan 🔒

Splits the range into multiple smaller ranges. It assumes the input ranges list is created by group_ranges_for_seq_scan().