mito2
Module region

Mito region.

Modules

opener 🔒

Region opener.

options

Options for a region.

version 🔒

Version control of mito engine.

Structs

ManifestContext 🔒

Context to update the region manifest.

ManifestStats 🔒

Manifest stats.

MitoRegion 🔒

Metadata and runtime status of a region.

OpeningRegions 🔒

Opening regions

RegionMap 🔒

Regions indexed by ids.

RegionUsage

Region status include region id, memtable usage, sst usage, wal usage and manifest usage.

Enums

RegionLeaderState

RegionRoleState

Constants

ESTIMATED_WAL_FACTOR 🔒

This is the approximate factor to estimate the size of wal.

Type Aliases

ManifestContextRef 🔒

MitoRegionRef 🔒

OpeningRegionsRef 🔒

RegionMapRef 🔒