mito2::region

Module opener

Region opener.

Structs

RegionLoadCacheTask 🔒

A task to load and fill the region file cache.

RegionOpener 🔒

Builder to create a new MitoRegion or open an existing one.

Traits

PartitionExprFetcher

A fetcher to retrieve partition expr for a region.

Functions

can_load_cache 🔒

check_recovered_region 🔒

Checks whether the recovered region has the same schema as region to create.

get_object_store

Returns an object store corresponding to name. If name is None, this method returns the default object store.

maybe_load_cache 🔒

Loads all index (Puffin) files from the version into the write cache.

replay_memtable 🔒

Replays the mutations from WAL and inserts mutations to memtable of given region.

sanitize_region_options 🔒

Updates region options by persistent options.

version_builder_from_manifest 🔒

Creates a version builder from a region manifest.

Type Aliases

PartitionExprFetcherRef