Module mito2::region::options

Options for a region.

If we add options in this mod, we also need to modify store_api::mito_engine_options.

Modules

prefix_inverted_index 🔒

prefix_partition_tree 🔒

prefix_twcs 🔒

Structs

IndexOptions

Options for index.

InvertedIndexOptions

Options for the inverted index.

PartitionTreeOptions

Partition tree memtable options.

RegionOptions

Options that affect the entire region.

RegionOptionsWithoutEnum 🔒

We need to define a new struct without enum fields as #[serde(default)] does not support external tagging.

TwcsOptions

Time window compaction options.

Enums

CompactionOptions

Options for compactions

MemtableOptions

Options for region level memtable.

Constants

DEFAULT_INDEX_SEGMENT_ROW_COUNT 🔒

Functions

deserialize_ignore_column_ids 🔒

options_map_to_value 🔒

Converts the options map to a json object.

validate_enum_options 🔒

Validates whether the options_map has valid options for specific enum_tag_key and returns true if the map contains enum options.