mito2::region

Module utils

Structs

RegionFileCopier

A copier for copying files from a region to another region.

RegionMetadataLoader

A loader for loading metadata from a region dir.

Enums

FileDescriptor

A descriptor for a file.

Functions

build_copy_file_paths 🔒

Builds the source and target file paths for a given file descriptor.

build_delete_file_path 🔒