mito2

Module remap_manifest

Re-exports

pub use crate::error::Error;

pub use crate::error::Result;

Structs

RemapManifest

Remaps file references from old region manifests to new region manifests.

RemapResult

Result of manifest remapping, including new manifests and statistics.

RemapStats

Statistical information about the manifest remapping.