mito2
Module row_converter

Modules

dense 🔒

sparse 🔒

Structs

DensePrimaryKeyCodec

A memory-comparable row [Value] encoder/decoder.

SortField

SparsePrimaryKeyCodec

A codec for sparse key of metrics. It requires the input primary key columns are sorted by the column name in lexicographical order. It encodes the column id of the physical region.

SparseValues

Sparse values representation.

Enums

CompositeValues

Composite values decoded from primary key bytes.

Constants

COLUMN_ID_ENCODE_SIZE

The size of the column id in the encoded sparse row.

Traits

PrimaryKeyCodec

PrimaryKeyCodecExt

Row value encoder/decoder.

PrimaryKeyFilter

Functions

build_primary_key_codec

Builds a primary key codec from region metadata.

build_primary_key_codec_with_fields

Builds a primary key codec from region metadata.