mito2::schedule::remote_job_scheduler

Trait RemoteJobScheduler

pub trait RemoteJobScheduler:
    Send
    + Sync
    + 'static {
    // Required method
    fn schedule<'life0, 'async_trait>(
        &'life0 self,
        job: RemoteJob,
        notifier: Box<dyn Notifier>,
    ) -> Pin<Box<dyn Future<Output = Result<JobId, RemoteJobSchedulerError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
}

RemoteJobScheduler is a trait that defines the API to schedule remote jobs. For example, a compaction job can be scheduled remotely as the following workflow:

participant User participant MitoEngine participant CompactionScheduler participant Plugins participant RemoteJobScheduler User->>MitoEngine: Initiates compaction MitoEngine->>CompactionScheduler: schedule_compaction() CompactionScheduler->>Plugins: Handle plugins CompactionScheduler->>RemoteJobScheduler: schedule(CompactionJob) RemoteJobScheduler-->>CompactionScheduler: Returns Job UUID CompactionScheduler-->>MitoEngine: Task scheduled with Job UUID MitoEngine-->>User: Compaction task scheduled

Required Methods

fn schedule<'life0, 'async_trait>( &'life0 self, job: RemoteJob, notifier: Box<dyn Notifier>, ) -> Pin<Box<dyn Future<Output = Result<JobId, RemoteJobSchedulerError>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Sends a job to the scheduler and returns a UUID for the job.