mito2::sst

Module index

Modules

bloom_filter 🔒

fulltext_index 🔒

indexer 🔒

intermediate

inverted_index 🔒

puffin_manager

statistics 🔒

store 🔒

Structs

IndexBaseOutput

Base output of the index creation.

IndexBuildScheduler

IndexBuildStatus

Tracks the index build status of a region scheduled by the IndexBuildScheduler.

IndexBuildTask

IndexOutput

Output of the index creation.

Indexer

The index creator that hides the error handling details.

IndexerBuilderImpl 🔒

Enums

IndexBuildOutcome

Outcome of an index build task.

IndexBuildType

Type of an index build task.

Constants

TYPE_BLOOM_FILTER_INDEX 🔒

TYPE_FULLTEXT_INDEX 🔒

TYPE_INVERTED_INDEX 🔒

Traits

IndexerBuilder

Functions

decode_primary_keys_with_counts 🔒

Decodes primary keys from a flat format RecordBatch. Returns a list of (decoded_pk_value, count) tuples where count is the number of occurrences.

trigger_index_background_download 🔒

Triggers background download of an index file to the local cache.

Type Aliases

BloomFilterOutput

Output of the bloom filter creation.

FulltextIndexOutput

Output of the fulltext index creation.

InvertedIndexOutput

Output of the inverted index creation.

ResultMpscSender

Mpsc output result sender.