mito2::sst
Module parquet

SST in parquet format.

Modules

file_range 🔒

Structs and functions for reading ranges from a parquet file. A file range is usually a row group in a parquet file.

format

Format to store in parquet.

helper 🔒

metadata 🔒

page_reader 🔒

Parquet page reader.

reader

Parquet reader.

row_group

Ports private structs from parquet crate.

row_selection 🔒

stats 🔒

Statistics of parquet SSTs.

writer

Parquet writer.

Structs

SstInfo

Parquet SST info returned by the writer.

WriteOptions

Parquet write options.

Constants

DEFAULT_READ_BATCH_SIZE 🔒

Default batch size to read parquet files.

DEFAULT_ROW_GROUP_SIZE 🔒

Default row group size for parquet files.

PARQUET_METADATA_KEY

Key of metadata in parquet SST.