Module mito2::sst::parquet::format

Format to store in parquet.

We store three internal columns in parquet:

__primary_key, the primary key of the row (tags). Type: dictionary(uint16, binary)

__sequence, the sequence number of a row. Type: uint64

__op_type, the op type of the row. Type: uint8

The schema of a parquet file is:

field 0, field 1, ..., field N, time index, primary key, sequence, op type

We stores fields in the same order as RegionMetadata::field_columns().

Structs

ReadFormat 🔒

Helper for reading the SST format.

WriteFormat 🔒

Helper for writing the SST format.

Constants

FIXED_POS_COLUMN_NUM 🔒

Number of columns that have fixed positions.

Functions

internal_fields 🔒

Fields for internal columns.

new_primary_key_array 🔒

Creates a new array for specific primary_key.

primary_key_offsets 🔒

Compute offsets of different primary keys in the array.

to_sst_arrow_schema 🔒

Gets the arrow schema to store in parquet.