mito2::sst::parquet

Module reader

Parquet reader.

Macros

handle_index_error 🔒

Structs

ParquetReader

Parquet batch reader to read our SST format.

ParquetReaderBuilder

Parquet SST reader builder.

ReaderFilterMetrics 🔒

Metrics of filtering rows groups and rows.

ReaderMetrics 🔒

Parquet reader metrics.

RowGroupReaderBase 🔒

Reader to read a row group of a parquet file.

RowGroupReaderBuilder 🔒

Builder to build a [ParquetRecordBatchReader] for a row group.

SimpleFilterContext 🔒

Context to evaluate the column filter for a parquet file.

Enums

MaybeFilter 🔒

The filter to evaluate or the prune result of the default value.

ReaderState 🔒

The state of a ParquetReader.

Constants

INDEX_TYPE_BLOOM 🔒

INDEX_TYPE_FULLTEXT 🔒

INDEX_TYPE_INVERTED 🔒

Traits

RowGroupReaderContext 🔒

RowGroupReaderContext represents the fields that cannot be shared between different RowGroupReaders.

Functions

apply_selection_and_update_metrics 🔒

pruned_by_default 🔒

Prune a column by its default value. Returns false if we can’t create the default value or evaluate the filter.

Type Aliases

RowGroupReader 🔒

RowGroupReader that reads from [FileRange].