mito2::sst::parquet

Module row_group

Ports private structs from parquet crate.

Structs

ColumnChunkIterator 🔒

Implements [PageIterator] for a single column chunk, yielding a single [PageReader]

InMemoryRowGroup

An in-memory collection of column chunks

ParquetFetchMetrics

Metrics for tracking page/row group fetch operations.

ParquetFetchMetricsData

Inner data for ParquetFetchMetrics.

RowGroupBase 🔒

Enums

ColumnChunkData 🔒

An in-memory column chunk

Functions

align_to_pooled_buf_size 🔒

Aligns the given size to the multiple of the pooled buffer size.

compute_total_range_size 🔒

Computes the max possible buffer size to read the given ranges. Returns (aligned_size, unaligned_size) where: