mito2::sst::parquet

Module writer

Parquet writer.

Structs

ObjectStoreWriterFactory

ParquetWriter

Parquet SST writer.

SizeAwareWriter 🔒

Workaround for [AsyncArrowWriter] does not provide a method to get total bytes written after close.

SourceStats 🔒

Traits

WriterFactory

Functions

timestamp_range_from_array 🔒

Gets min and max timestamp from an timestamp array.