mito2::worker

Module handle_drop

Handling drop request.

Constants

GC_TASK_INTERVAL_SEC 🔒

MAX_RETRY_TIMES 🔒

Functions

later_drop_task_with_global_gc 🔒

later_drop_task_without_global_gc 🔒

Background GC task to remove the entire region path once one of the following conditions is true:

remove_region_dir_once 🔒

Removes region dir if there is no parquet files, returns whether the directory is removed. If force = true, always removes the dir.

remove_region_with_retry 🔒