mito2::worker::handle_drop

Function later_drop_task

async fn later_drop_task(
    region_id: RegionId,
    region_path: String,
    object_store: ObjectStore,
    dropping_regions: Arc<RegionMap>,
    gc_duration: Duration,
) -> bool

Background GC task to remove the entire region path once one of the following conditions is true:

It finds there is no parquet file left.

After gc_duration.

Returns whether the path is removed.

This task will retry on failure and keep running until finished. Any resource captured by it will not be released before then. Be sure to only pass weak reference if something is depended on ref-count mechanism.