mito2

Module access_layer

Structs

AccessLayer

A layer to access SST files under the same directory.

RegionFilePathFactory 🔒

Path provider that builds paths in region storage path.

SstWriteRequest

Contents to build a SST.

TempFileCleaner 🔒

Cleaner to remove temp files on the atomic write dir.

WriteCachePathProvider 🔒

Path provider that builds paths in local write cache.

Enums

OperationType

OperationType represents the origin of the SstWriteRequest.

Constants

ATOMIC_WRITE_DIR

OLD_ATOMIC_WRITE_DIR

For compatibility. Remove this after a major version release.

Traits

FilePathProvider

Path provider for SST file and index file.

Functions

clean_dir 🔒

Clean the directory.

new_fs_cache_store 🔒

Type Aliases

AccessLayerRef

SstInfoArray

SST write results.