mito2
Module cache

Cache for the engine.

Modules

cache_size 🔒

Cache size of different cache value.

file_cache 🔒

A cache for files.

index 🔒

write_cache 🔒

A write-through cache for remote object stores.

Structs

CacheManager

Manages cached data for the engine.

CacheManagerBuilder

Builder to construct a CacheManager.

ColumnPagePath

Path to column pages in the SST file.

PageValue

Cached row group pages for a column.

SelectorResultKey

Cache key for time series row selector result.

SelectorResultValue

Cached result for time series row selector.

SstMetaKey 🔒

Cache key (region id, file id) for SST meta.

Enums

CacheStrategy

Cache strategies that may only enable a subset of caches.

PageKey

Cache key for pages of a SST row group.

Constants

FILE_TYPE 🔒

Metrics type key for files on the local store.

PAGE_TYPE 🔒

Metrics type key for pages.

SELECTOR_RESULT_TYPE 🔒

Metrics type key for selector result cache.

SST_META_TYPE 🔒

Metrics type key for sst meta.

VECTOR_TYPE 🔒

Metrics type key for vector.

Functions

meta_cache_weight 🔒

page_cache_weight 🔒

selector_result_cache_hit

Increases selector cache hit metrics.

selector_result_cache_miss

Increases selector cache miss metrics.

selector_result_cache_weight 🔒

update_hit_miss 🔒

Updates cache hit/miss metrics.

vector_cache_weight 🔒

Type Aliases

CacheManagerRef

PageCache 🔒

Maps (region, file, row group, column) to PageValue.

SelectorResultCache 🔒

Maps (file id, row group id, time series row selector) to SelectorResultValue.

SstMetaCache 🔒

Maps (region id, file id) to [ParquetMetaData].

VectorCache 🔒

Maps [Value] to a vector that holds this value repeatedly.