mito2::cache

Module file_cache

A cache for files.

Structs

FileCache 🔒

A file cache manages files on local store and evict files based on size.

IndexKey 🔒

Key of file cache index.

IndexValue 🔒

An entity that describes the file in the file cache.

Enums

FileType

Type of the file.

Constants

FILE_DIR 🔒

Subdirectory of cached files for write.

Functions

cache_file_path 🔒

Generates the path to the cached file.

parse_index_key 🔒

Parse index key from the file name.

Type Aliases

FileCacheRef 🔒