mito2
Module compaction

Modules

buckets 🔒

compactor

picker

run 🔒

This file contains code to find sorted runs in a set if ranged items and along with the best way to merge these items to satisfy the desired run count.

task 🔒

twcs 🔒

window 🔒

Structs

CompactionOutput

CompactionRequest

Region compaction request.

CompactionScheduler 🔒

Compaction scheduler tracks and manages compaction tasks.

CompactionSstReaderBuilder 🔒

Builders to create BoxedBatchReader for compaction.

CompactionStatus 🔒

Status of running and pending region compaction tasks.

PendingCompaction 🔒

Pending compaction request that is supposed to run after current task is finished, typically used for manual compactions.

SerializedCompactionOutput

SerializedCompactionOutput is a serialized version of CompactionOutput by replacing FileHandle with FileMeta.

Functions

find_ttl 🔒

Finds TTL of table by first examine table options then database options.

get_expired_ssts 🔒

Finds all expired SSTs across levels.

time_range_to_predicate 🔒

Converts time range to predicates so that rows outside the range will be filtered.

ts_to_lit 🔒