mito2::compaction

Module compactor

Structs

CompactionRegion

CompactionRegion represents a region that needs to be compacted. It’s the subset of MitoRegion.

CompactionVersion

Region version for compaction that does not hold memtables.

DefaultCompactor

DefaultCompactor is the default implementation of Compactor.

MergeOutput

[MergeOutput] represents the output of merging SST files.

OpenCompactionRegionRequest

OpenCompactionRegionRequest represents the request to open a compaction region.

Traits

Compactor

Compactor is the trait that defines the compaction logic.

Functions

open_compaction_region

Open a compaction region from a compaction request. It’s simple version of RegionOpener::open().