mito2::compaction

Module twcs

Structs

TwcsPicker

TwcsPicker picks files of which the max timestamp are in the same time window as compaction candidates.

Window 🔒

Constants

LEVEL_COMPACTED 🔒

Functions

assign_to_windows 🔒

Assigns files to windows with predefined window size (in seconds) by their max timestamps.

find_latest_window_in_seconds 🔒

Finds the latest active writing window among all files. Returns None when there are no files or all files are corrupted.

log_pick_result 🔒