Module mito2::compaction::window

Structs

WindowedCompactionPicker

Compaction picker that splits the time range of all involved files to windows, and merges the data segments intersects with those windows of files together so that the output files never overlaps.

Functions

assign_files_to_time_windows 🔒

Assigns files to time windows. If file does not contain a time range in metadata, it will be assigned to a special bucket i64::MAX (normally no timestamp can be aligned to this bucket) so that all files without timestamp can be compacted together.

build_output 🔒

file_time_bucket_span 🔒

Calculates timestamp span between start and end timestamp.