mito2::flush

Trait WriteBufferManager

pub trait WriteBufferManager:
    Send
    + Sync
    + Debug {
    // Required methods
    fn should_flush_engine(&self) -> bool;
    fn should_stall(&self) -> bool;
    fn reserve_mem(&self, mem: usize);
    fn schedule_free_mem(&self, mem: usize);
    fn free_mem(&self, mem: usize);
    fn memory_usage(&self) -> usize;
}

Global write buffer (memtable) manager.

Tracks write buffer (memtable) usages and decide whether the engine needs to flush.

Required Methods

fn should_flush_engine(&self) -> bool

Returns whether to trigger the engine.

fn should_stall(&self) -> bool

Returns whether to stall write requests.

fn reserve_mem(&self, mem: usize)

Reserves mem bytes.

fn schedule_free_mem(&self, mem: usize)

Tells the manager we are freeing mem bytes.

We are in the process of freeing mem bytes, so it is not considered when checking the soft limit.

fn free_mem(&self, mem: usize)

We have freed mem bytes.

fn memory_usage(&self) -> usize

Returns the total memory used by memtables.