Module mito2::manifest::storage

Structs

CHECKPOINT_RE 🔒

CheckpointMetadata 🔒

DELTA_RE 🔒

ManifestObjectStore

Enums

FileKey 🔒

Key to identify a manifest file.

Constants

DEFAULT_MANIFEST_COMPRESSION_TYPE 🔒

FALL_BACK_COMPRESS_TYPE 🔒

Due to backward compatibility, it is possible that the user’s manifest file has not been compressed. So when we encounter problems, we need to fall back to FALL_BACK_COMPRESS_TYPE for processing.

FETCH_MANIFEST_PARALLELISM 🔒

LAST_CHECKPOINT_FILE 🔒

Functions

checkpoint_checksum 🔒

checkpoint_file

delta_file

file_compress_type

Return’s the file compress algorithm by file extension.

file_version

Return’s the file manifest version from path

gen_path

is_checkpoint_file

is_delta_file

manifest_compress_type

Returns the [CompressionType] according to whether to compress manifest files.

verify_checksum 🔒