mito2

Module memtable

Memtables are write buffers for regions.

Re-exports

pub use bulk::part::EncodedBulkPart;

pub use bulk::part::BulkPart;

pub use bulk::part::BulkPartEncoder;

pub use bulk::part::BulkPartMeta;

pub use bulk::part::UnorderedPart;

pub use bulk::part::record_batch_estimated_size;

pub use bulk::part::sort_primary_key_record_batch;

Modules

builder 🔒

Builders for time-series memtable

bulk

Memtable implementation for bulk load

partition_tree

Memtable implementation based on a partition tree.

simple_bulk_memtable

stats 🔒

Internal metrics of the memtable.

time_partition

Partitions memtables by time.

time_series

version 🔒

Memtable version.

Structs

AllocTracker

Memtable memory allocation tracker.

EncodedRange

Encoded range in the memtable.

KeyValues

Key value view of a mutation.

MemScanMetrics

Metrics for scanning a memtable.

MemScanMetricsData 🔒

MemtableBuilderProvider 🔒

Provider of memtable builders for regions.

MemtableRange

A range in the memtable.

MemtableRangeContext

Context shared by ranges of the same memtable.

MemtableRanges

Ranges in a memtable.

MemtableStats

RangesOptions

Options for querying ranges from a memtable.

Enums

MemtableConfig

Config for memtables.

Traits

IterBuilder

Builder to build an iterator to read the range. The builder should know the projection and the predicate to build the iterator.

Memtable

In memory write buffer.

MemtableBuilder

Builder to build a new Memtable.

Type Aliases

BoxedBatchIterator

BoxedIterBuilder

BoxedRecordBatchIterator

MemtableBuilderRef

MemtableId

Id for memtables.

MemtableRangeContextRef

MemtableRef