mito2

Module memtable

Memtables are write buffers for regions.

Re-exports

pub use crate::memtable::key_values::KeyValues;

Modules

builder 🔒

Builders for time-series memtable

bulk

Memtable implementation for bulk load

key_values

partition_tree

Memtable implementation based on a partition tree.

simple_bulk_memtable 🔒

stats 🔒

Internal metrics of the memtable.

time_partition

Partitions memtables by time.

time_series

version 🔒

Memtable version.

Structs

AllocTracker

Memtable memory allocation tracker.

EncodedBulkPart

MemtableBuilderProvider 🔒

Provider of memtable builders for regions.

MemtableRange

A range in the memtable.

MemtableRangeContext

Context shared by ranges of the same memtable.

MemtableRanges

Ranges in a memtable.

MemtableStats

Enums

MemtableConfig

Config for memtables.

Traits

IterBuilder

Builder to build an iterator to read the range. The builder should know the projection and the predicate to build the iterator.

Memtable

In memory write buffer.

MemtableBuilder

Builder to build a new Memtable.

Type Aliases

BoxedBatchIterator

BoxedIterBuilder

MemtableBuilderRef

MemtableId

Id for memtables.

MemtableRangeContextRef

MemtableRef