mito2::memtable

Module bulk

Memtable implementation for bulk load

Modules

context

Context for iterating bulk memtable.

part

Bulk part encoder/decoder.

part_reader

row_group_reader 🔒

Structs

BulkMemtable

Memtable that ingests and scans parts directly.

BulkMemtableBuilder

Builder to build a BulkMemtable.

BulkPartWrapper 🔒

BulkParts 🔒

All parts in a bulk memtable.

BulkRangeIterBuilder

Iterator builder for bulk range

CompactDispatcher

Scheduler to run compact tasks in background.

EncodedBulkRangeIterBuilder 🔒

Iterator builder for encoded bulk range

EncodedPartWrapper 🔒

MemCompactTask 🔒

A memtable compact task to run in background.

MemtableCompactor 🔒

MergingFlagsGuard 🔒

RAII guard for managing merging flags. Automatically resets merging flags when dropped if the merge operation wasn’t successful.

Enums

PartToMerge 🔒

Enum to wrap different types of parts for unified merging.