mito2::memtable::partition_tree::merger

Trait Node

pub trait Node: Ord {
    // Required methods
    fn is_valid(&self) -> bool;
    fn is_behind(&self, other: &Self) -> bool;
    fn advance(&mut self, len: usize) -> Result<()>;
    fn current_item_len(&self) -> usize;
    fn search_key_in_current_item(&self, other: &Self) -> Result<usize, usize>;
}

Nodes of merger’s heap.

Required Methods

fn is_valid(&self) -> bool

Returns true if current node is not exhausted.

fn is_behind(&self, other: &Self) -> bool

Whether the other node is behind (exclusive) current node.

fn advance(&mut self, len: usize) -> Result<()>

Advances len rows from current batch. If current batch is empty it fetches next batch from the node.

Panics

If the node is invalid.

fn current_item_len(&self) -> usize

Length of current item.

fn search_key_in_current_item(&self, other: &Self) -> Result<usize, usize>

Searches first key of other in current item and returns the index.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.