mito2::memtable

Trait IterBuilder

pub trait IterBuilder: Send + Sync {
    // Required method
    fn build(
        &self,
        metrics: Option<MemScanMetrics>,
    ) -> Result<BoxedBatchIterator>;

    // Provided methods
    fn is_record_batch(&self) -> bool { ... }
    fn build_record_batch(
        &self,
        metrics: Option<MemScanMetrics>,
    ) -> Result<BoxedRecordBatchIterator> { ... }
    fn encoded_range(&self) -> Option<EncodedRange> { ... }
}

Builder to build an iterator to read the range. The builder should know the projection and the predicate to build the iterator.

Required Methods

fn build(&self, metrics: Option<MemScanMetrics>) -> Result<BoxedBatchIterator>

Returns the iterator to read the range.

Provided Methods

fn is_record_batch(&self) -> bool

Returns whether the iterator is a record batch iterator.

fn build_record_batch( &self, metrics: Option<MemScanMetrics>, ) -> Result<BoxedRecordBatchIterator>

Returns the record batch iterator to read the range.

fn encoded_range(&self) -> Option<EncodedRange>

Returns the EncodedRange if the range is already encoded into SST.