mito2::memtable

Trait Memtable

pub trait Memtable:
    Send
    + Sync
    + Debug {
    // Required methods
    fn id(&self) -> MemtableId;
    fn write(&self, kvs: &KeyValues) -> Result<()>;
    fn write_one(&self, key_value: KeyValue<'_>) -> Result<()>;
    fn write_bulk(&self, part: BulkPart) -> Result<()>;
    fn ranges(
        &self,
        projection: Option<&[ColumnId]>,
        predicate: PredicateGroup,
        sequence: Option<SequenceNumber>,
    ) -> Result<MemtableRanges>;
    fn is_empty(&self) -> bool;
    fn freeze(&self) -> Result<()>;
    fn stats(&self) -> MemtableStats;
    fn fork(&self, id: MemtableId, metadata: &RegionMetadataRef) -> MemtableRef;
}

In memory write buffer.

Required Methods

fn id(&self) -> MemtableId

Returns the id of this memtable.

fn write(&self, kvs: &KeyValues) -> Result<()>

Writes key values into the memtable.

fn write_one(&self, key_value: KeyValue<'_>) -> Result<()>

Writes one key value pair into the memtable.

fn write_bulk(&self, part: BulkPart) -> Result<()>

Writes an encoded batch of into memtable.

fn ranges( &self, projection: Option<&[ColumnId]>, predicate: PredicateGroup, sequence: Option<SequenceNumber>, ) -> Result<MemtableRanges>

Returns the ranges in the memtable. The returned map contains the range id and the range after applying the predicate.

fn is_empty(&self) -> bool

Returns true if the memtable is empty.

fn freeze(&self) -> Result<()>

Turns a mutable memtable into an immutable memtable.

fn stats(&self) -> MemtableStats

Returns the MemtableStats info of Memtable.

fn fork(&self, id: MemtableId, metadata: &RegionMetadataRef) -> MemtableRef

Forks this (immutable) memtable and returns a new mutable memtable with specific memtable id.

A region must freeze the memtable before invoking this method.