mito2

Module read

Common structs and utilities for reading data.

Modules

compat

Utilities to adapt readers with different schema.

dedup

Utilities to remove duplicate rows from a sorted batch.

last_row

Utilities to read the last row of each time series.

merge

Merge reader implementation.

plain_batch

Batch without an encoded primary key.

projection

Utilities for projection.

prune 🔒

range 🔒

Structs for partition ranges.

scan_region 🔒

Scans a region according to the scan request.

scan_util 🔒

Utilities for scanners.

seq_scan 🔒

Sequential scan.

series_scan 🔒

Per-series scan implementation.

unordered_scan 🔒

Unordered scanner.

Structs

Batch

Storage internal representation of a batch of rows for a primary key (time series).

BatchBuilder

Builder to build Batch.

BatchChecker 🔒

A struct to check the batch is monotonic.

BatchColumn

A column in a Batch.

ScannerMetrics 🔒

Local metrics for scanners.

Enums

Source

Async Batch reader and iterator wrapper.

Constants

TIMESTAMP_KEY_LEN 🔒

Len of timestamp in arrow row format.

Traits

BatchReader

Async batch reader.

Functions

concat_arrays 🔒

Helper function to concat arrays from iter.

Type Aliases

BoxedBatchReader

Pointer to BatchReader.

BoxedBatchStream

Pointer to a stream that yields Batch.