mito2::read

Module compat

Utilities to adapt readers with different schema.

Structs

CompatBatch 🔒

A helper struct to adapt schema of the batch to an expected schema.

CompatFields 🔒

Helper to make fields compatible.

CompatPrimaryKey 🔒

Helper to make primary key compatible.

CompatReader

Reader to adapt schema of underlying reader to expected schema.

RewritePrimaryKey 🔒

Adapter to rewrite primary key.

Enums

IndexOrDefault 🔒

Index in source batch or a default value to fill a column.

Functions

has_same_columns_and_pk_encoding 🔒

Returns true if left and right have same columns and primary key encoding.

may_compat_fields 🔒

Creates a CompatFields if needed.

may_compat_primary_key 🔒

Creates a CompatPrimaryKey if needed.

may_rewrite_primary_key 🔒