mito2::read

Module last_row

Utilities to read the last row of each time series.

Structs

LastRowCacheReader 🔒

Last row reader that returns the cached last rows for row group.

LastRowReader 🔒

Reader to keep the last row for each time series. It assumes that batches from the input reader are

LastRowSelector

Common struct that selects only the last row of each time series.

RowGroupLastRowReader 🔒

Enums

RowGroupLastRowCachedReader 🔒

Cached last row reader for specific row group. If the last rows for current row group are already cached, this reader returns the cached value. If cache misses, RowGroupLastRowReader reads last rows from row group and updates the cache upon finish.

Functions

push_yielded_batches 🔒

Push last row into yielded_batches.