mito2::read

Module range

Structs for partition ranges.

Structs

FileRangeBuilder 🔒

Builder to create file ranges.

MemRangeBuilder 🔒

Builder to create mem ranges.

RangeBuilderList 🔒

List to manages the builders to create file ranges. Each scan partition should have its own list. Mutex inside this list is used to allow moving the list to different streams in the same partition.

RangeMeta 🔒

Meta data of a partition range. If the scanner is [UnorderedScan], each meta only has one row group or memtable. If the scanner is [SeqScan], each meta may have multiple row groups and memtables.

RowGroupIndex 🔒

Index to access a row group.

SourceIndex 🔒

Index and metadata for a memtable or file.

Constants

ALL_ROW_GROUPS 🔒

Functions

group_ranges_for_seq_scan 🔒

Groups ranges by time range. It assumes each input range only contains a file or a memtable.

maybe_split_ranges_for_seq_scan 🔒

Splits the range into multiple smaller ranges. It assumes the input ranges list is created by group_ranges_for_seq_scan().