mito2::read

Module series_scan

Per-series scan implementation.

Structs

SenderList 🔒

List of senders.

SeriesBatch 🔒

Batches of the same series.

SeriesDistributor 🔒

The distributor scans series and distributes them to different partitions.

SeriesScan

Scans a region and returns sorted rows of a series in the same partition.

Constants

SEND_TIMEOUT 🔒

Timeout to send a batch to a sender.

Functions

new_channel_list 🔒

new_partition_metrics 🔒

Type Aliases

ReceiverList 🔒

List of receivers.