mito2::row_converter
Trait RowCodec

pub trait RowCodec {
    // Required methods
    fn encode<'a, I>(&self, row: I) -> Result<Vec<u8>>
       where I: Iterator<Item = ValueRef<'a>>;
    fn encode_to_vec<'a, I>(&self, row: I, buffer: &mut Vec<u8>) -> Result<()>
       where I: Iterator<Item = ValueRef<'a>>;
    fn decode(&self, bytes: &[u8]) -> Result<Vec<Value>>;
}

Row value encoder/decoder.

Required Methods

fn encode<'a, I>(&self, row: I) -> Result<Vec<u8>>
where I: Iterator<Item = ValueRef<'a>>,

Encodes rows to bytes.

Note

Ensure the length of row iterator matches the length of fields.

fn encode_to_vec<'a, I>(&self, row: I, buffer: &mut Vec<u8>) -> Result<()>
where I: Iterator<Item = ValueRef<'a>>,

Encodes rows to specific vec.

Note

Ensure the length of row iterator matches the length of fields.

fn decode(&self, bytes: &[u8]) -> Result<Vec<Value>>

Decode row values from bytes.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.