mito2::schedule

Module remote_job_scheduler

Structs

CompactionJob

CompactionJob is a remote job that compacts a set of files in a compaction service.

CompactionJobResult

CompactionJobResult is the result of a compaction job.

DefaultNotifier 🔒

DefaultNotifier is a default implementation of Notifier that sends WorkerRequest to the mito engine.

JobId

Unique id for a remote job.

RemoteJobSchedulerError

RemoteJobSchedulerSnafu 🔒

SNAFU context selector for the RemoteJobSchedulerError error

Enums

RemoteJob

RemoteJob is a job that can be executed remotely. For example, a remote compaction job.

RemoteJobResult

RemoteJobResult is the result of a remote job.

Traits

Notifier

Notifier is used to notify the mito engine when a remote job is completed.

RemoteJobScheduler

RemoteJobScheduler is a trait that defines the API to schedule remote jobs. For example, a compaction job can be scheduled remotely as the following workflow:

Type Aliases

RemoteJobSchedulerRef