mito2

Module sst

Sorted strings tables.

Modules

file

Structures to describe metadata of files.

file_purger

file_ref

index

location

parquet

SST in parquet format.

version 🔒

SST version.

Structs

FlatSchemaOptions

Options of flat schema.

SeriesEstimator 🔒

Gets the estimated number of series from record batches.

Enums

FormatType

Format type of the SST file.

Constants

DEFAULT_WRITE_BUFFER_SIZE

Default write buffer size, it should be greater than the default minimum upload part of S3 (5mb).

DEFAULT_WRITE_CONCURRENCY

Default number of concurrent write, it only works on object store backend(e.g., S3).

Functions

flat_sst_arrow_schema_column_num

Returns the number of columns in the flat format.

internal_fields 🔒

Fields for internal columns.

plain_internal_fields 🔒

Fields for internal columns.

tag_maybe_to_dictionary_field 🔒

Helper function to create a dictionary field from a field if it is a string column.

to_dictionary_field 🔒

Helper function to create a dictionary field from a field.

to_flat_sst_arrow_schema

Gets the arrow schema to store in parquet.

to_plain_sst_arrow_schema

Gets the arrow schema to store in parquet.

to_sst_arrow_schema

Gets the arrow schema to store in parquet.