mito2
Module sst

Sorted strings tables.

Modules

file

Structures to describe metadata of files.

file_purger

index

location

parquet

SST in parquet format.

version 🔒

SST version.

Constants

DEFAULT_WRITE_BUFFER_SIZE

Default write buffer size, it should be greater than the default minimum upload part of S3 (5mb).

DEFAULT_WRITE_CONCURRENCY

Default number of concurrent write, it only works on object store backend(e.g., S3).

Functions

internal_fields 🔒

Fields for internal columns.

to_sst_arrow_schema

Gets the arrow schema to store in parquet.