mito2::sst::parquet

Module file_range

Structs and functions for reading ranges from a parquet file. A file range is usually a row group in a parquet file.

Structs

FileRange

A range of a parquet SST. Now it is a row group. We can read different file ranges in parallel.

FileRangeContext 🔒

Context shared by ranges of the same parquet SST.

RangeBase 🔒

Common fields for a range to read and filter batches.

Type Aliases

FileRangeContextRef 🔒