Module mito2::sst::parquet::row_selection

Functions

add_or_merge_selector 🔒

Helper function to either add a new RowSelector to selectors or merge it with the last one if they are of the same type (both skip or both select).

row_selection_from_row_ranges 🔒

Converts an iterator of row ranges into a RowSelection by creating a sequence of RowSelectors. Returns the RowSelection and the number of rows that were skipped.

Type Aliases

SkipRowCount 🔒