mito2::sst::parquet

Module row_selection

Structs

RowGroupSelection

A selection of row groups.

RowSelectionWithCount 🔒

A row selection with its count.

Functions

add_or_merge_selector 🔒

Helper function to either add a new RowSelector to selectors or merge it with the last one if they are of the same type (both skip or both select).

row_selection_from_row_ranges 🔒

Converts an iterator of row ranges into a RowSelection by creating a sequence of RowSelectors.

row_selection_from_sorted_row_ids 🔒

Converts an iterator of sorted row IDs into a RowSelection.

selector_len 🔒

Returns the length of the selectors in the selection.