mito2::time_provider

Trait TimeProvider

pub trait TimeProvider:
    Debug
    + Send
    + Sync {
    // Required methods
    fn current_time_millis(&self) -> i64;
    fn elapsed_since(&self, current_millis: i64) -> i64;

    // Provided method
    fn wait_duration(&self, duration: Duration) -> Duration { ... }
}

Trait to get current time and deal with durations.

We define the trait to simplify time related tests.

Required Methods

fn current_time_millis(&self) -> i64

Returns current time in millis.

fn elapsed_since(&self, current_millis: i64) -> i64

Returns millis elapsed since specify time.

Provided Methods

fn wait_duration(&self, duration: Duration) -> Duration

Computes the actual duration to wait from an expected one.